requirements.txt
sentence-transformers~=3.3.1
torch
numpy~=2.0.2
Flask~=3.1.0
requests~=2.28.1
beautifulsoup4~=4.11.1
lxml
python-dotenv~=1.0.1

openai~=1.56.0